dart transpiler: trim input lines
