+++ tests/algorithms/x/Dart/physics/speeds_of_gas_molecules.dart
@@ -1,27 +1,4 @@
 // Generated by Mochi transpiler
-import 'dart:convert';
-import 'dart:io';
-
-int _nowSeed = 0;
-bool _nowSeeded = false;
-void _initNow() {
-  var s = Platform.environment['MOCHI_NOW_SEED'];
-  if (s != null && s.isNotEmpty) {
-    var v = int.tryParse(s);
-    if (v != null) {
-      _nowSeed = v;
-      _nowSeeded = true;
-    }
-  }
-}
-int _now() {
-  if (_nowSeeded) {
-    _nowSeed = (_nowSeed * 1664525 + 1013904223) % 2147483647;
-    return _nowSeed;
-  }
-  return DateTime.now().microsecondsSinceEpoch;
-}
-
 dynamic _substr(dynamic s, num start, num end) {
   int n = s.length;
   int s0 = start.toInt();
@@ -86,21 +63,8 @@ double mps_speed_of_molecule(double temperature, double molar_mass) {
 }
 
 void main() {
-  var _benchMem0 = ProcessInfo.currentRss;
-  var _benchSw = Stopwatch()..start();
-  _initNow();
-  {
-  var _benchMem0 = ProcessInfo.currentRss;
-  var _benchSw = Stopwatch()..start();
   print(_str(avg_speed_of_molecule(273.0, 0.028)));
   print(_str(avg_speed_of_molecule(300.0, 0.032)));
   print(_str(mps_speed_of_molecule(273.0, 0.028)));
   print(_str(mps_speed_of_molecule(300.0, 0.032)));
-  _benchSw.stop();
-  var _benchMem1 = ProcessInfo.currentRss;
-  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
-}
-  _benchSw.stop();
-  var _benchMem1 = ProcessInfo.currentRss;
-  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
 }
