+++ tests/algorithms/x/Dart/project_euler/problem_006/sol2.dart
@@ -1,29 +1,6 @@
 // Generated by Mochi transpiler
-import 'dart:convert';
-import 'dart:io';
-
-int _nowSeed = 0;
-bool _nowSeeded = false;
-void _initNow() {
-  var s = Platform.environment['MOCHI_NOW_SEED'];
-  if (s != null && s.isNotEmpty) {
-    var v = int.tryParse(s);
-    if (v != null) {
-      _nowSeed = v;
-      _nowSeeded = true;
-    }
-  }
-}
-int _now() {
-  if (_nowSeeded) {
-    _nowSeed = (_nowSeed * 1664525 + 1013904223) % 2147483647;
-    return _nowSeed;
-  }
-  return DateTime.now().microsecondsSinceEpoch;
-}
-
-String _substr(String s, num start, num end) {
-  var n = s.length;
+dynamic _substr(dynamic s, num start, num end) {
+  int n = s.length;
   int s0 = start.toInt();
   int e0 = end.toInt();
   if (s0 < 0) s0 += n;
@@ -33,10 +10,13 @@ String _substr(String s, num start, num end) {
   if (e0 < 0) e0 = 0;
   if (e0 > n) e0 = n;
   if (s0 > e0) s0 = e0;
-  return s.substring(s0, e0);
+  if (s is String) {
+    return s.substring(s0, e0);
+  }
+  return s.sublist(s0, e0);
 }
 
-String _str(dynamic v) { if (v is double && v == v.roundToDouble()) { return v.toInt().toString(); } return v.toString(); }
+String _str(dynamic v) { if (v is double && v == v.roundToDouble()) { var i = v.toInt(); if (i == 0) return '0'; return i.toString(); } return v.toString(); }
 
 int solution(int n) {
   int sum = n * (n + 1) ~/ 2;
@@ -46,18 +26,5 @@ int solution(int n) {
 }
 
 void main() {
-  var _benchMem0 = ProcessInfo.currentRss;
-  var _benchSw = Stopwatch()..start();
-  _initNow();
-  {
-  var _benchMem0 = ProcessInfo.currentRss;
-  var _benchSw = Stopwatch()..start();
   print("solution() = " + _str(solution(100)));
-  _benchSw.stop();
-  var _benchMem1 = ProcessInfo.currentRss;
-  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
-}
-  _benchSw.stop();
-  var _benchMem1 = ProcessInfo.currentRss;
-  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
 }
