+++ tests/algorithms/x/Dart/project_euler/problem_008/sol2.dart
@@ -1,29 +1,6 @@
 // Generated by Mochi transpiler
-import 'dart:convert';
-import 'dart:io';
-
-int _nowSeed = 0;
-bool _nowSeeded = false;
-void _initNow() {
-  var s = Platform.environment['MOCHI_NOW_SEED'];
-  if (s != null && s.isNotEmpty) {
-    var v = int.tryParse(s);
-    if (v != null) {
-      _nowSeed = v;
-      _nowSeeded = true;
-    }
-  }
-}
-int _now() {
-  if (_nowSeeded) {
-    _nowSeed = (_nowSeed * 1664525 + 1013904223) % 2147483647;
-    return _nowSeed;
-  }
-  return DateTime.now().microsecondsSinceEpoch;
-}
-
-String _substr(String s, num start, num end) {
-  var n = s.length;
+dynamic _substr(dynamic s, num start, num end) {
+  int n = s.length;
   int s0 = start.toInt();
   int e0 = end.toInt();
   if (s0 < 0) s0 += n;
@@ -33,7 +10,10 @@ String _substr(String s, num start, num end) {
   if (e0 < 0) e0 = 0;
   if (e0 > n) e0 = n;
   if (s0 > e0) s0 = e0;
-  return s.substring(s0, e0);
+  if (s is String) {
+    return s.substring(s0, e0);
+  }
+  return s.sublist(s0, e0);
 }
 
 String N = "73167176531330624919225119674426574742355349194934" + "96983520312774506326239578318016984801869478851843" + "85861560789112949495459501737958331952853208805511" + "12540698747158523863050715693290963295227443043557" + "66896648950445244523161731856403098711121722383113" + "62229893423380308135336276614282806444486645238749" + "30358907296290491560440772390713810515859307960866" + "70172427121883998797908792274921901699720888093776" + "65727333001053367881220235421809751254540594752243" + "52584907711670556013604839586446706324415722155397" + "53697817977846174064955149290862569321978468622482" + "83972241375657056057490261407972968652414535100474" + "82166370484403199890008895243450658541227588666881" + "16427171479924442928230863465674813919123162824586" + "17866458359124566529476545682848912883142607690042" + "24219022671055626321111109370544217506941658960408" + "07198403850962455444362981230987879927244284909188" + "84580156166097919133875499200524063689912560717606" + "05886116467109405077541002256983155200055935729725" + "71636269561882670428252483600823257530420752963450";
@@ -44,7 +24,7 @@ int solution(String n) {
     int product = 1;
     int j = 0;
     while (j < 13) {
-    product = product * ((n.substring(i + j, i + j + 1)).codeUnitAt(0));
+    product = product * (int.parse(_substr(n, i + j, i + j + 1)));
     j = j + 1;
   }
     if (product > max_product) {
@@ -56,21 +36,8 @@ int solution(String n) {
 }
 
 void main() {
-  var _benchMem0 = ProcessInfo.currentRss;
-  var _benchSw = Stopwatch()..start();
-  _initNow();
-  {
-  var _benchMem0 = ProcessInfo.currentRss;
-  var _benchSw = Stopwatch()..start();
   print(solution("13978431290823798458352374"));
   print(solution("13978431295823798458352374"));
   print(solution("1397843129582379841238352374"));
   print(solution(N));
-  _benchSw.stop();
-  var _benchMem1 = ProcessInfo.currentRss;
-  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
-}
-  _benchSw.stop();
-  var _benchMem1 = ProcessInfo.currentRss;
-  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
 }
