+++ tests/algorithms/x/Dart/project_euler/problem_010/sol3.dart
@@ -1,6 +1,6 @@
 // Generated by Mochi transpiler
-String _substr(String s, num start, num end) {
-  var n = s.length;
+dynamic _substr(dynamic s, num start, num end) {
+  int n = s.length;
   int s0 = start.toInt();
   int e0 = end.toInt();
   if (s0 < 0) s0 += n;
@@ -10,10 +10,13 @@ String _substr(String s, num start, num end) {
   if (e0 < 0) e0 = 0;
   if (e0 > n) e0 = n;
   if (s0 > e0) s0 = e0;
-  return s.substring(s0, e0);
+  if (s is String) {
+    return s.substring(s0, e0);
+  }
+  return s.sublist(s0, e0);
 }
 
-String _str(dynamic v) { if (v is double && v == v.roundToDouble()) { return v.toInt().toString(); } return v.toString(); }
+String _str(dynamic v) { if (v is double && v == v.roundToDouble()) { var i = v.toInt(); if (i == 0) return '0'; return i.toString(); } return v.toString(); }
 
 int isqrt(int n) {
   int r = 0;
@@ -27,7 +30,7 @@ int solution(int n) {
   List<bool> sieve = <bool>[];
   int i = 0;
   while (i <= n) {
-    sieve = (sieve..add(false));
+    sieve = [...sieve, false];
     i = i + 1;
   }
   while (sieve.length <= 0) { sieve.add(false); } sieve[0] = true;
