+++ tests/algorithms/x/Dart/project_euler/problem_014/sol2.dart
@@ -1,8 +1,29 @@
 // Generated by Mochi transpiler
+import 'dart:convert';
 import 'dart:io';
 
-String _substr(String s, num start, num end) {
-  var n = s.length;
+int _nowSeed = 0;
+bool _nowSeeded = false;
+void _initNow() {
+  var s = Platform.environment['MOCHI_NOW_SEED'];
+  if (s != null && s.isNotEmpty) {
+    var v = int.tryParse(s);
+    if (v != null) {
+      _nowSeed = v;
+      _nowSeeded = true;
+    }
+  }
+}
+int _now() {
+  if (_nowSeeded) {
+    _nowSeed = (_nowSeed * 1664525 + 1013904223) % 2147483647;
+    return _nowSeed;
+  }
+  return DateTime.now().microsecondsSinceEpoch;
+}
+
+dynamic _substr(dynamic s, num start, num end) {
+  int n = s.length;
   int s0 = start.toInt();
   int e0 = end.toInt();
   if (s0 < 0) s0 += n;
@@ -12,7 +33,10 @@ String _substr(String s, num start, num end) {
   if (e0 < 0) e0 = 0;
   if (e0 > n) e0 = n;
   if (s0 > e0) s0 = e0;
-  return s.substring(s0, e0);
+  if (s is String) {
+    return s.substring(s0, e0);
+  }
+  return s.sublist(s0, e0);
 }
 
 Map<int, int> collatz_cache = {1: 1};
@@ -20,7 +44,7 @@ int collatz_length(int n) {
   int _num = n;
   List<int> sequence = <int>[];
   while (!collatz_cache.containsKey(_num)) {
-    sequence = (sequence..add(_num));
+    sequence = [...sequence, _num];
     if (_num % 2 == 0) {
     _num = _num ~/ 2 as int;
   } else {
@@ -52,8 +76,21 @@ int solution(int limit) {
   return max_start;
 }
 
-String input_str = stdin.readLineSync() ?? '';
+String input_str = stdin.readLineSync()?.trim() ?? '';
 int limit = int.parse(input_str);
 void main() {
+  var _benchMem0 = ProcessInfo.currentRss;
+  var _benchSw = Stopwatch()..start();
+  _initNow();
+  {
+  var _benchMem0 = ProcessInfo.currentRss;
+  var _benchSw = Stopwatch()..start();
   print(solution(limit));
+  _benchSw.stop();
+  var _benchMem1 = ProcessInfo.currentRss;
+  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
+}
+  _benchSw.stop();
+  var _benchMem1 = ProcessInfo.currentRss;
+  print(jsonEncode({"duration_us": _benchSw.elapsedMicroseconds, "memory_bytes": (_benchMem1 - _benchMem0).abs(), "name": "main"}));
 }
