+++ tests/algorithms/x/Dart/project_euler/problem_014/sol2.out
@@ -1,12 +1,1 @@
-Unhandled exception:
-FormatException: Invalid number (at character 1)
-
-^
-
-#0      int._handleFormatError (dart:core-patch/integers_patch.dart:135:7)
-#1      int.parse (dart:core-patch/integers_patch.dart:53:14)
-#2      limit (file:///workspace/mochi/tests/algorithms/x/Dart/project_euler/problem_014/sol2.dart:56:17)
-#3      limit (file:///workspace/mochi/tests/algorithms/x/Dart/project_euler/problem_014/sol2.dart)
-#4      main (file:///workspace/mochi/tests/algorithms/x/Dart/project_euler/problem_014/sol2.dart:58:18)
-#5      _delayEntrypointInvocation.<anonymous closure> (dart:isolate-patch/isolate_patch.dart:314:19)
-#6      _RawReceivePort._handleMessage (dart:isolate-patch/isolate_patch.dart:193:12)
+97
